CORE-8643 Refactor ArgumentGroupContentPanelAppearance into appearance interface and default appearance class in themes package
