+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/apps/Applications.gwt.xml
@@ -104,6 +104,10 @@
         <when-type-is class="org.iplantc.de.apps.integration.client.view.tools.ToolSearchField.ToolSearchFieldAppearance"/>
     </replace-with>
 
+    <replace-with class="org.iplantc.de.theme.base.client.apps.integration.ArgumentGroupContentPanelDefaultAppearance">
+        <when-type-is class="org.iplantc.de.apps.integration.client.view.widgets.AppTemplateContentPanel.ArgumentGroupContentPanelAppearance"/>
+    </replace-with>
+
     <!--Launch Widget-->
     <replace-with class="org.iplantc.de.theme.base.client.apps.widgets.AppLaunchViewDefaultAppearance">
         <when-type-is class="org.iplantc.de.apps.widgets.client.view.AppLaunchView.AppLaunchViewAppearance"/>
